refactor: move to utilities folder
